fix(sc-21663): review round — measured audio decomposition, candle bar recalibration, cross-check tool, comment corrections
Review blockers:
- Audio 2.3 parity widening now carries its measured per-op decomposition
  instead of the (wrong) 2x/layer amplification story: port-vs-oracle
  deviations are flat kernel noise (1e-4..7.5e-4 per op — pmetal-fork vs
  wheel Metal GEMMs genuinely differ at 1e-4-class, not 1e-7) through
  block 6; the whole 8.086e-2 appears at the closing per-row RMS-norm,
  which rescales rows spanning a 272x norm range (video: 134x) and turns
  absolute kernel noise into relative error on near-cancelled register
  rows (worst row is a register row; corr(log rowRMS, log rowErr)=-0.62).
  Not a defect: stock torch differs from the oracle by 6.4e-2 on the same
  tensor. Derivation committed in the test docs, with the 2.5-numbers
  reconciliation (register-row cancellation is weights+inputs dependent).
- candle 2.3 parity: docstring's oracle/dtype claims rewritten; bars
  recalibrated as documented PROVISIONAL derived bounds with a valid/global
  split (video 3e-2/1.5e-1, audio 5e-2/2.5e-1); first CUDA run records
  actuals and tightens.
- dump_ltx_te_golden.py now applies all THREE overrides (rope patch was
  missing) with hasattr guards.
- The false 'rope delta alone holds the parity gate at 1.3e-2/8.8e-2'
  comments replaced with the verified-rebuild mutation measurement
  (f64 revert: video 2.005e-3 / audio 8.238e-2, still green; 'Compiling'
  line confirmed on every mutation run).

Majors/minors: candle 2.5 twin's stale RED narrative updated with the
bf16-activation caveat; candle connector carries the bf16-deficiency
breadcrumb (6.094e-2 vs 6e-2 measured on mlx, decision rides with the
first CUDA run); the two bf16/f32 measurement pairs are now labeled with
their provenance (isolated-connector vs full-encoder) at every citation;
--cross-check [--old-fixture] committed on the golden dumper (reproduces
stock-torch 4.658e-2/6.364e-2 vs the fixture and 1.877e-4/1.716e-4 for
the sigma+erf variant vs the pre-fix fixture); theta read from self;
hasattr asserts; scaled positions all-f32 op order (value-neutral for
power-of-two max_pos — parity numbers byte-identical); gelu_tanh
ExactEpilogues bit-identity noted. Committed fixture unchanged.

+++ crates/media/candle-gen/candle-gen-ltx/src/rope.rs
@@ -86,8 +86,12 @@ pub fn precompute_connector_freqs(
     };
     // f64 exponentials rounded to f32 BEFORE the position multiply — exactly upstream ltx_core's
     // `generate_freq_grid_np` (its "double precision" covers only the log-spaced grid; the angles
-    // are formed in f32). Keeping f64 through cos/sin perturbs the top frequencies by ~1e-3 rad,
-    // which the connector's 2·sigmoid gates amplify coherently (sc-21663; mirrors the mlx port).
+    // are formed in f32). The table matches the reference's to 1 f32 ULP. Keeping f64 through
+    // cos/sin (the old behaviour, inherited from mlx_video) perturbs the top frequencies by up to
+    // ~9.4e-4 — a real bit-level infidelity, though the mlx-side verified-rebuild mutation run
+    // measured only an ~e-3-scale end-to-end parity shift from it (per-op deviations do not
+    // compound across blocks — sc-21663; see mlx connector_parity.rs's decomposition). Kept for
+    // faithfulness, pinned by construction rather than by a tolerance bar.
     let indices: Vec<f32> = (0..num_indices)
         .map(|i| (theta.powf(i as f64 * step) * (PI / 2.0)) as f32)
         .collect();
@@ -100,8 +104,9 @@ pub fn precompute_connector_freqs(
     let mut cos_out = vec![0f32; total];
     let mut sin_out = vec![0f32; total];
     for t in 0..seq {
-        // position = raw index t, scaled by max_pos, *2-1 (mlx connector `rope`).
-        let scaled = (t as f64 / max_pos as f64 * 2.0 - 1.0) as f32;
+        // All-f32 op order, as upstream's `get_fractional_positions` computes it (bit-identical
+        // to f64-then-round for the shipped power-of-two max_pos, exact for any max_pos this way).
+        let scaled = (t as f32 / max_pos as f32) * 2.0 - 1.0;
         for h in 0..num_heads {
             for p in 0..head_half {
                 let flat = h * head_half + p;
